map_graph/requirements.txt
pyxel>=2.0.0
pyyaml>=6.0
